backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Web & Networking
websockets==12.0
httpx==0.25.2
aiofiles==23.2.1

# Data & Storage
redis==5.0.4
aioredis==2.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Optional Dependencies
# Uncomment as needed

# Advanced AI Models
transformers==4.35.2
sentence-transformers==2.2.2
chromadb==0.4.18
langchain==0.0.350

# Database Drivers
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0          # MongoDB

# Message Queues
celery==5.3.4
# kombu==5.3.4

# Caching
# aiocache==0.12.2

# Security
# cryptography==41.0.8
# authlib==1.2.1

# File Processing
# pillow==10.1.0
# pypdf2==3.0.1

# Monitoring & Observability
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0

# Production WSGI/ASGI
# gunicorn==21.2.0
# hypercorn==0.15.0

numpy<2.0
psutil
